add mute and exit

+++ Client/core/src/de/bremen/screens/BaseScreen.java
@@ -6,8 +6,8 @@ import de.bremen.MainClient;
 
 public abstract class BaseScreen implements Screen {
     protected MainClient game;
-    public static final int WIDTH = 1920;
-    public static final int HEIGHT = 1080;
+    public static final float WIDTH = 1920;
+    public static final float HEIGHT = 1080;
 
 
     public BaseScreen(MainClient game) {
